+++ Assets/GraphicsLabor/Scripts/Attributes/LaborerAttributes/LabelAttribute.cs
@@ -1,0 +1,14 @@
+﻿using System;
+
+namespace GraphicsLabor.Scripts.Attributes.LaborerAttributes
+{
+    public class LabelAttribute: Attribute, ILaborerAttribute
+    {
+        public readonly string _label;
+
+        public LabelAttribute(string label)
+        {
+            _label = label;
+        }
+    }
+}