+++ Assets/GraphicsLabor/Scripts/Attributes/LaborerAttributes/ShowMessageAttribute.cs
@@ -8,16 +8,16 @@ namespace GraphicsLabor.Scripts.Attributes.LaborerAttributes
     /// Allows to display a box above the field to give a warning or info
     /// </summary>
     [AttributeUsage(AttributeTargets.Field, AllowMultiple = true, Inherited = true)]
-    public class ShowMessage : PropertyAttribute, ILaborerAttribute
+    public class ShowMessageAttribute : PropertyAttribute, ILaborerAttribute
     {
 
         public string Message { get; private set; }
         public MessageType MessageType { get; private set; }
 
-        public ShowMessage(string message, MessageType messageType = MessageType.None)
+        public ShowMessageAttribute(string message, MessageType messageType = MessageType.None)
         {
-            Message = message;
-            MessageType = messageType;
+            this.Message = message;
+            this.MessageType = messageType;
         }
     }