+++ Assets/GraphicsLabor/Scripts/Attributes/LaborerAttributes/ShowPropertyAttribute.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace GraphicsLabor.Scripts.Attributes.LaborerAttributes
+{
+    [AttributeUsage(AttributeTargets.Property, AllowMultiple = false, Inherited = true)]
+    public class ShowPropertyAttribute : Attribute, ILaborerAttribute
+    {
+        
+    }
+}