+++ Assets/GraphicsLabor/Scripts/Attributes/Utility/Enums.cs
@@ -1,0 +1,9 @@
+﻿namespace GraphicsLabor.Scripts.Attributes.Utility
+{
+    public enum ConditionOperator
+    {
+        // Maybe do a NOR if possible
+        And,
+        Or
+    }
+}