+++ Assets/GraphicsLabor/Scripts/Editor/Drawers/ShowMessageDrawer.cs
@@ -1,37 +1,34 @@
 ﻿using GraphicsLabor.Scripts.Attributes.LaborerAttributes;
-using GraphicsLabor.Scripts.Editor.CustomInspector;
+using GraphicsLabor.Scripts.Editor.Utility;
 using UnityEditor;
 using UnityEngine;
 
 namespace GraphicsLabor.Scripts.Editor.Drawers
 {
-    [CustomPropertyDrawer(typeof(ShowMessage))]
+    [CustomPropertyDrawer(typeof(ShowMessageAttribute))]
     public class ShowMessageDrawer : DecoratorDrawer {
     
-        public override float GetHeight()
-        {
-            return GetHelpBoxHeight();
-        }
+        public override float GetHeight() => GetHelpBoxHeight();
         
         public override void OnGUI(Rect rect)
         {
-            ShowMessage info = attribute as ShowMessage;
+            if (attribute is not ShowMessageAttribute showMessageAttribute) return;
 
             float indent = LaborerEditorGUI.GetIndentLength(rect);
-            Rect infoBoxRect = new Rect(
+            Rect infoBoxRect = new(
                 rect.x + indent,
                 rect.y,
                 rect.width - indent,
                 GetHelpBoxHeight());
             
-            EditorGUI.HelpBox(infoBoxRect  , info?.Message, info?.MessageType ?? MessageType.None);
+            EditorGUI.HelpBox(infoBoxRect  , showMessageAttribute.Message, showMessageAttribute.MessageType);
         }
         
         private float GetHelpBoxHeight()
         {
-            ShowMessage infoMessageAttribute = (ShowMessage)attribute;
+            ShowMessageAttribute showMessageAttributeAttribute = (ShowMessageAttribute)attribute;
             float minHeight = EditorGUIUtility.singleLineHeight * 2.0f;
-            float desiredHeight = GUI.skin.box.CalcHeight(new GUIContent(infoMessageAttribute.Message), EditorGUIUtility.currentViewWidth);
+            float desiredHeight = GUI.skin.box.CalcHeight(new GUIContent(showMessageAttributeAttribute.Message), EditorGUIUtility.currentViewWidth);
             float height = Mathf.Max(minHeight, desiredHeight);
 
             return height;