+++ Assets/GraphicsLabor/Scripts/Editor/Utility/LaborerEditorGUI.cs
@@ -1,0 +1,264 @@
+using System;
+using System.Collections;
+using System.Linq;
+using System.Reflection;
+using GraphicsLabor.Scripts.Attributes.LaborerAttributes;
+using UnityEditor;
+using UnityEditor.SceneManagement;
+using UnityEngine;
+using UnityEngine.SceneManagement;
+using Object = UnityEngine.Object;
+
+namespace GraphicsLabor.Scripts.Editor.Utility
+{
+    public static class LaborerEditorGUI
+    {
+        private static readonly GUIStyle ButtonStyle = new GUIStyle(GUI.skin.button) { richText = true };
+
+        public static void Button(Object target, MethodInfo methodInfo)
+        {
+            if (methodInfo.GetParameters().All(p => p.IsOptional))
+            {
+                ButtonAttribute buttonAttribute =
+                    (ButtonAttribute)methodInfo.GetCustomAttributes(typeof(ButtonAttribute), true)[0];
+                string buttonText = string.IsNullOrEmpty(buttonAttribute.Text)
+                    ? ObjectNames.NicifyVariableName(methodInfo.Name)
+                    : buttonAttribute.Text;
+
+                EButtonEnableMode mode = buttonAttribute.SelectedEnableMode;
+                bool buttonEnabled = true;
+                buttonEnabled &=
+                    mode == EButtonEnableMode.Always ||
+                    mode == EButtonEnableMode.Editor && !Application.isPlaying ||
+                    mode == EButtonEnableMode.Playmode && Application.isPlaying;
+
+                bool methodIsCoroutine = methodInfo.ReturnType == typeof(IEnumerator);
+                if (methodIsCoroutine)
+                {
+                    buttonEnabled &= Application.isPlaying;
+                }
+
+                EditorGUI.BeginDisabledGroup(!buttonEnabled);
+
+                if (GUILayout.Button(buttonText, ButtonStyle))
+                {
+                    object[] defaultParams = methodInfo.GetParameters().Select(p => p.DefaultValue).ToArray();
+                    IEnumerator methodResult = (IEnumerator)methodInfo.Invoke(target, defaultParams);
+
+                    if (!Application.isPlaying)
+                    {
+                        // Set target object and scene dirty to serialize changes to disk
+                        // If not it just goes into the void basically
+                        EditorUtility.SetDirty(target);
+
+                        PrefabStage stage = PrefabStageUtility.GetCurrentPrefabStage();
+                        if (stage != null)
+                        {
+                            // Prefab mode
+                            EditorSceneManager.MarkSceneDirty(stage.scene);
+                        }
+                        else
+                        {
+                            // Normal scene
+                            EditorSceneManager.MarkSceneDirty(SceneManager.GetActiveScene());
+                        }
+                    }
+                    else if (methodResult != null && target is MonoBehaviour behaviour)
+                    {
+                        behaviour.StartCoroutine(methodResult);
+                    }
+                }
+
+                EditorGUI.EndDisabledGroup();
+            }
+            else
+            {
+                const string warning = nameof(ButtonAttribute) + " works only on methods with no parameters";
+                EditorGUILayout.HelpBox(warning, MessageType.Warning);
+            }
+        }
+        
+        public static void PropertyField(Rect rect, SerializedProperty property, bool includeChildren)
+        {
+            // Check if visible
+            if (!PropertyUtility.IsVisible(property)) return;
+            
+            bool enabled = PropertyUtility.IsEnabled(property);
+
+            using (new EditorGUI.DisabledScope(disabled: !enabled))
+            {
+                EditorGUI.PropertyField(rect, property, PropertyUtility.GetLabel(property), includeChildren);
+            }
+        }
+
+        public static void LayoutField(SerializedProperty property)
+        {
+            if (!PropertyUtility.IsVisible(property)) return;
+            
+            bool isEnabled = PropertyUtility.IsEnabled(property);
+                
+            using (new EditorGUI.DisabledScope(disabled: !isEnabled))
+            {
+                EditorGUILayout.PropertyField(property, PropertyUtility.GetLabel(property), true);
+            }
+        }
+        
+        public static void LayoutProperty(Object target, PropertyInfo property)
+        {
+            object value = property.GetValue(target, null);
+
+            if (value == null)
+            {
+                string warning =
+                    $"{ObjectNames.NicifyVariableName(property.Name)} is null. {nameof(ShowPropertyAttribute)} doesn't support reference types with null value";
+                EditorGUILayout.HelpBox(warning, MessageType.Warning);
+            }
+            else if (!DrawField(value, ObjectNames.NicifyVariableName(property.Name)))
+            {
+                string warning = $"{nameof(ShowPropertyAttribute)} doesn't support {property.PropertyType.Name} types";
+                EditorGUILayout.HelpBox(warning, MessageType.Warning);
+            }
+        }
+        
+        public static bool DrawField(object value, string label)
+        {
+            using (new EditorGUI.DisabledScope(disabled: true))
+            {
+                bool isDrawn = true;
+                Type valueType = value.GetType();
+
+                if (valueType == typeof(bool))
+                {
+                    EditorGUILayout.Toggle(label, (bool)value);
+                }
+                else if (valueType == typeof(short))
+                {
+                    EditorGUILayout.IntField(label, (short)value);
+                }
+                else if (valueType == typeof(ushort))
+                {
+                    EditorGUILayout.IntField(label, (ushort)value);
+                }
+                else if (valueType == typeof(int))
+                {
+                    EditorGUILayout.IntField(label, (int)value);
+                }
+                else if (valueType == typeof(uint))
+                {
+                    EditorGUILayout.LongField(label, (uint)value);
+                }
+                else if (valueType == typeof(long))
+                {
+                    EditorGUILayout.LongField(label, (long)value);
+                }
+                else if (valueType == typeof(ulong))
+                {
+                    EditorGUILayout.TextField(label, ((ulong)value).ToString());
+                }
+                else if (valueType == typeof(float))
+                {
+                    EditorGUILayout.FloatField(label, (float)value);
+                }
+                else if (valueType == typeof(double))
+                {
+                    EditorGUILayout.DoubleField(label, (double)value);
+                }
+                else if (valueType == typeof(string))
+                {
+                    EditorGUILayout.TextField(label, (string)value);
+                }
+                else if (valueType == typeof(Vector2))
+                {
+                    EditorGUILayout.Vector2Field(label, (Vector2)value);
+                }
+                else if (valueType == typeof(Vector3))
+                {
+                    EditorGUILayout.Vector3Field(label, (Vector3)value);
+                }
+                else if (valueType == typeof(Vector4))
+                {
+                    EditorGUILayout.Vector4Field(label, (Vector4)value);
+                }
+                else if (valueType == typeof(Vector2Int))
+                {
+                    EditorGUILayout.Vector2IntField(label, (Vector2Int)value);
+                }
+                else if (valueType == typeof(Vector3Int))
+                {
+                    EditorGUILayout.Vector3IntField(label, (Vector3Int)value);
+                }
+                else if (valueType == typeof(Color))
+                {
+                    EditorGUILayout.ColorField(label, (Color)value);
+                }
+                else if (valueType == typeof(Bounds))
+                {
+                    EditorGUILayout.BoundsField(label, (Bounds)value);
+                }
+                else if (valueType == typeof(Rect))
+                {
+                    EditorGUILayout.RectField(label, (Rect)value);
+                }
+                else if (valueType == typeof(RectInt))
+                {
+                    EditorGUILayout.RectIntField(label, (RectInt)value);
+                }
+                else if (typeof(Object).IsAssignableFrom(valueType))
+                {
+                    EditorGUILayout.ObjectField(label, (Object)value, valueType, true);
+                }
+                else if (valueType.BaseType == typeof(Enum))
+                {
+                    EditorGUILayout.EnumPopup(label, (Enum)value);
+                }
+                else if (valueType.BaseType == typeof(TypeInfo))
+                {
+                    EditorGUILayout.TextField(label, value.ToString());
+                }
+                else
+                {
+                    isDrawn = false;
+                }
+
+                return isDrawn;
+            }
+        }
+        
+        public static void HorizontalLine(Rect rect, float height, Color color)
+        {
+            rect.height = height;
+            EditorGUI.DrawRect(rect, color);
+        }
+        
+        public static void HelpBox(Rect rect, string message, MessageType type, UnityEngine.Object context = null, bool logToConsole = false)
+        {
+            EditorGUI.HelpBox(rect, message, type);
+
+            if (!logToConsole) return;
+            
+            switch (type)
+            {
+                case MessageType.None:
+                case MessageType.Info:
+                    Debug.Log(message, context);
+                    break;
+                case MessageType.Warning:
+                    Debug.LogWarning(message, context);
+                    break;
+                case MessageType.Error:
+                    Debug.LogError(message, context);
+                    break;
+                default:
+                    throw new ArgumentOutOfRangeException(nameof(type), type, null);
+            }
+        }
+
+        public static float GetIndentLength(Rect sourceRect)
+        {
+            Rect indentRect = EditorGUI.IndentedRect(sourceRect);
+            float indentLength = indentRect.x - sourceRect.x;
+
+            return indentLength;
+        }
+    }
+}