+++ CHANGELOG.md
@@ -5,6 +5,39 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+
+## [2.0.0] - 2023-10-29
+
+### Added
+
+- ShowIf Attribute
+- HideIf Attribute
+- Expandable Attribute for ScriptableObjects
+- HorizontalSeparator Decoration Attribute
+- ColorExtension for HorizontalSeparator 
+- ShowProperty Attribute
+- Scene Attribute
+- Drawers for Expandable, Scene and BasePropertyDrawer
+- TestScene For Testing
+- A TestScriptableObject to test Attributes
+- Documentation on certain Functions, more will be made on future patches
+
+
+### Changed
+
+- ReadOnly Attribute now is usable with other attributes and handled by the GraphicsLaborInspector
+- ReadOnly Attribute can no longer change the display name of the field
+- TestScrip to show usage of Attributes
+
+
+### Removed
+
+- ReadOnlyDrawer (not necessary anymore)
+- CHANGELOG and README meta files
+
+
+
 ## [1.1.1] - 2023-10-29
 
 ### Added
@@ -16,6 +49,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - License from MIT to GNU
 - DrawMode enum to LaborerDrawMode to avoid name conflicts
+
 
 
 ## [1.1.0] - 2023-10-29